Main

See merge request thebaapcompany/baap-academic-service!139

+++ routes/classes.routes.js
@@ -8,16 +8,22 @@ const TokenService = require("../services/token.services");
 
 router.post(
     "/",
-    checkSchema(require("../dto/classes.dto")), TokenService.checkPermission(["EMDC2"]),
+    checkSchema(require("../dto/classes.dto")),
+    TokenService.checkPermission(["EMDC2"]),
     async (req, res, next) => {
         if (ValidationHelper.requestValidationErrors(req, res)) {
             return;
         }
-        const existingRecord = await service.getByCourseIdAndGroupId(req.body.groupId, req.body.name, req.body.courseId);
+        const existingRecord = await service.getByCourseIdAndGroupId(
+            req.body.groupId,
+            req.body.name,
+            req.body.courseId
+        );
         console.log(existingRecord);
         if (existingRecord.data) {
-
-            return res.status(400).json({ error: "Name With The Same GroupId Already Exists." });
+            return res
+                .status(400)
+                .json({ error: "Name With The Same GroupId Already Exists." });
         }
         const classId = +Date.now();
         req.body.classId = classId;
@@ -26,37 +32,63 @@ router.post(
     }
 );
 
-router.get("/all", TokenService.checkPermission(["EMDC1"]), async (req, res) => {
-    const serviceResponse = await service.getAllByCriteria(req.query);
-    requestResponsehelper.sendResponse(res, serviceResponse);
-});
+router.get(
+    "/all",
+    TokenService.checkPermission(["EMDC1"]),
+    async (req, res) => {
+        const serviceResponse = await service.getAllByCriteria(req.query);
+        requestResponsehelper.sendResponse(res, serviceResponse);
+    }
+);
 
-router.delete("/:id", TokenService.checkPermission(["EMDC4"]), async (req, res) => {
-    const serviceResponse = await service.deleteById(req.params.id);
-    requestResponsehelper.sendResponse(res, serviceResponse);
-});
+router.delete(
+    "/:id",
+    TokenService.checkPermission(["EMDC4"]),
+    async (req, res) => {
+        const serviceResponse = await service.deleteById(req.params.id);
+        requestResponsehelper.sendResponse(res, serviceResponse);
+    }
+);
 
-router.put("/:id", TokenService.checkPermission(["EMDC3"]), async (req, res) => {
-    const serviceResponse = await service.updateById(req.params.id, req.body);
-    requestResponsehelper.sendResponse(res, serviceResponse);
-});
+router.put(
+    "/:id",
+    TokenService.checkPermission(["EMDC3"]),
+    async (req, res) => {
+        const serviceResponse = await service.updateById(
+            req.params.id,
+            req.body
+        );
+        requestResponsehelper.sendResponse(res, serviceResponse);
+    }
+);
 
-router.get("/:id", TokenService.checkPermission(["EMDC1"]), async (req, res) => {
-    const serviceResponse = await service.getById(req.params.id);
-    requestResponsehelper.sendResponse(res, serviceResponse);
-});
+router.get(
+    "/:id",
+    TokenService.checkPermission(["EMDC1"]),
+    async (req, res) => {
+        const serviceResponse = await service.getById(req.params.id);
+        requestResponsehelper.sendResponse(res, serviceResponse);
+    }
+);
 
-router.get("/all/getByGroupId/:groupId", TokenService.checkPermission(["EMDC1"]), async (req, res) => {
-    const groupId = req.params.groupId;
-    const criteria = {
-        classId: req.query.classId,
-        name: req.query.name,
-        courseId: req.query.courseId,
-        Department: req.query.departmentId
-    };
-    const serviceResponse = await service.getAllDataByGroupId(groupId, criteria);
-    requestResponsehelper.sendResponse(res, serviceResponse);
-});
+router.get(
+    "/all/getByGroupId/:groupId",
+    TokenService.checkPermission(["EMDC1"]),
+    async (req, res) => {
+        const groupId = req.params.groupId;
+        const criteria = {
+            classId: req.query.classId,
+            name: req.query.name,
+            courseId: req.query.courseId,
+            Department: req.query.departmentId,
+        };
+        const serviceResponse = await service.getAllDataByGroupId(
+            groupId,
+            criteria
+        );
+        requestResponsehelper.sendResponse(res, serviceResponse);
+    }
+);
 
 router.get("/getAllUsingLink/getByGroupId/:groupId", async (req, res) => {
     const groupId = req.params.groupId;
@@ -64,9 +96,12 @@ router.get("/getAllUsingLink/getByGroupId/:groupId", async (req, res) => {
         classId: req.query.classId,
         name: req.query.name,
         courseId: req.query.courseId,
-        Department: req.query.departmentId
+        Department: req.query.departmentId,
     };
-    const serviceResponse = await service.getAllDataByGroupId(groupId, criteria);
+    const serviceResponse = await service.getAllDataByGroupId(
+        groupId,
+        criteria
+    );
     requestResponsehelper.sendResponse(res, serviceResponse);
 });
 module.exports = router;
